Changed diameter computation for entity centered

+++ env_entite_centre.h
@@ -21,7 +21,7 @@ class Env_entite_centre : public Simulation {
         EspeceMoleculaire** especes;
 
     public :
-     Env_entite_centre(EspeceMoleculaire**, int, int);
+     Env_entite_centre(EspeceMoleculaire**, int, float);
      ~Env_entite_centre();
 
      void run();
